Add Runtime::generate_ruid for scrypto
