--- radix-engine-tests/tests/blueprints/transaction_runtime/src/transaction_runtime.rs
@@ -1,12 +1,0 @@
-use scrypto::prelude::*;
-
-#[blueprint]
-mod transaction_limits {
-    struct TransactionRuntimeTest {}
-
-    impl TransactionRuntimeTest {
-        pub fn get_transaction_hash() -> Hash {
-            Runtime::transaction_hash()
-        }
-    }
-}
