+++ radix-engine-tests/tests/core.rs
@@ -7,19 +7,6 @@ use scrypto_unit::*;
 use transaction::prelude::*;
 
 #[test]
-fn test_process_and_transaction() {
-    let mut test_runner = TestRunnerBuilder::new().build();
-    let package_address = test_runner.compile_and_publish("./tests/blueprints/core");
-
-    let manifest1 = ManifestBuilder::new()
-        .lock_fee_from_faucet()
-        .call_function(package_address, "RuntimeTest", "query", manifest_args![])
-        .build();
-    let receipt1 = test_runner.execute_manifest(manifest1, vec![]);
-    receipt1.expect_commit_success();
-}
-
-#[test]
 fn test_call() {
     let mut test_runner = TestRunnerBuilder::new().build();
     let (public_key, _, account) = test_runner.new_allocated_account();
